updated base64toBuffer to be in fileUtils
